requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
lxml
pytest-allure-adapter

